requirements.txt
python-dotenv==1.0.1
streamlit==1.40.0
openai==1.54.3
pdf2image==1.17.0